requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
botocore==1.34.50
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.6
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
julius==0.2.7
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.0.post0
lightning-utilities==0.10.1
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.1
openai-whisper
optuna==3.5.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.0.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==2.2.0.post0
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.10
yarl==1.9.4
yt-dlp==2023.12.30
